+++ src/Engine/UI/Widgets/Widget.h
@@ -51,7 +51,7 @@ namespace OE
 				virtual void OnMouseOver(){};
 				virtual void OnMouseOut(){};
 				virtual void OnMouseClick(){};
-				virtual void OnMouseMove(const int x, const int y){};
+				virtual void OnMouseMove(const double x, const double y){};
 				virtual void OnMouseButton(const int button, const int action){};
 				virtual void OnCharEvent(const int codepoint){};
 				virtual void OnKeyEvent(const int key, const int action, const int mods){};
