+++ src/Engine/UI/Windows/WindowManager.h
@@ -47,7 +47,7 @@ namespace OE
 
 				void OnMouseButton(const int key, const int action);
 				void OnMouseClick(const int startX, const int startY, const int endX, const int endY);
-				void OnMouseMove(const int x, const int y);
+				void OnMouseMove(const double x, const double y);
 				void OnKeyEvent(const int key, const int action, const int mods);
 				void OnCharEvent(const int codepoint);
